under construction

+++ script/sql/create_sql_from_model.php
@@ -1,15 +1,19 @@
 <?php
 /**
- * Copyright (c) 2017 Yohei Yoshikawa (https://github.com/yoo16/)
+ * Copyright (c) 2017 Yohei Yoshikawa 
  *
  */
 require_once dirname(__FILE__) . '/../../lib/Controller.php';
 
 echo('-- Create SQL --'.PHP_EOL);
-$pgsql = new PwEntity();
-$sql = $pgsql->createTablesSQLForProject();
+$pgsql = new PwPgsql();
+$pgsql->createTablesSQLForProject();
 
-if (file_exists(DB_DIR)) {
-    $sql_path = DB_DIR.'sql/create.sql';
-    file_put_contents($sql_path, $sql);
+if ($pgsql->sql_files && file_exists(DB_DIR)) {
+    foreach ($pgsql->sql_files as $file_name => $sql) {
+        if ($file_name) {
+            $sql_path = DB_DIR."sql/{$file_name}.sql";
+            file_put_contents($sql_path, $sql);
+        }
+    }
 }
